backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
python-dotenv
pytest
pydantic-settings
openai
pandas
numpy